+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Ordering.reversed, Ordering.nullsFirst, Ordering.nullsLast, and Ordering.lexicographical
   are now properties for consistency with the Iterable protocol.
+* Various optimization and improvements on the Range class.
 
 ## 1.7.2